Tidying up

+++ scripts/not_currently_used/01_analysis_wrapper.R
@@ -1,9 +1,10 @@
 # rm(list = ls())
 
-# path_to_dir <- "C:/Users/nicho/Documents/bauer_sync/projects/impc_mv_analysis/github_multivariate_phenotype/multivariate_phenotype"
-# renv::activate(path_to_dir)
-# renv::restore(path_to_dir)
+path_to_dir <- "C:/Users/nicho/Documents/bauer_sync/projects/impc_mv_analysis/github_multivariate_phenotype/multivariate_phenotype"
+renv::activate(path_to_dir)
+renv::restore(path_to_dir)
 arguments <- commandArgs()
+
 if("--args" %in% arguments){
   argnam.in <- data.frame(nam = c("run_type", "scen", "subsamseed"), 
                           coersion.fn = c("as.character", "as.numeric", "as.numeric"), 
@@ -12,8 +13,7 @@ if("--args" %in% arguments){
     assign(argnam.in$nam[i], eval(call(argnam.in$coersion.fn[i], arguments[grep("--args", arguments) + i])))
 } else {
   run_type <- c("demo", "main", "benchmark", "test_benchmark")[1]
-  scen <- 12
-  if(run_type == "main") scen <- 1
+  scen <- 1
   subsamseed <- 1
 }
 
@@ -47,13 +47,7 @@ N <- analysis_table$N[scen]
 P <- analysis_table$P[scen]
 nSig <- analysis_table$nSig[scen]
 Data <- analysis_table$Data[scen]
-MVphen_K <- analysis_table$MVphen_K[scen]
-n_subsamples <- analysis_table$n_subsamples[scen]
 XDmeth <- Meth
-
-file_list <- get_file_list(control = control, 
-                           file_core_name = analysis_table[scen, "file_core_name"], 
-                           subsamseed = subsamseed)
 
 Y_raw <- Data_all[[Data]]$Y_raw
 Y_zeroed <- Data_all[[Data]]$Y_zeroed
@@ -68,16 +62,25 @@ Y.eml$zero <- Data_all[[Data]]$Y_zeroed
 S.eml$raw <- Data_all[[Data]]$S_raw
 S.eml$zero <- Data_all[[Data]]$S_zeroed
 
+###################################
+# Choose phen subset if 
+ph_all <- colnames(Y_raw)
+if (run_type %in% c("demo", "test_benchmark")) {
+  phens_ok_to_use <- ph_all[colMeans(is.na(Y_raw[, ph_all])) < .8]
+  set.seed(1)
+  phens_to_use <- sample(x = phens_ok_to_use, size = P)
+} else {
+  phens_to_use <- ph_all
+}
 
 ###################################
 # Create and Load train-test samples information 
-train_test_list <- get_train_test_split(control = control, Data_all = Data_all, N = N, P = P, Data = Data, n_subsamples = n_subsamples)
+train_test_list <- get_train_test_split(control = control, N = N, P = P, Data = Data)
 train_test_list_file_curr <- file.path("output", "train_test_splits", paste0(Data, "_N_", N, "_P_", P, ".RDS"))
-# if (!file.exists(train_test_list_file_curr)) {
+if (!file.exists(train_test_list_file_curr)) {
   saveRDS(object = train_test_list, file = train_test_list_file_curr)
-# }
-phens_to_use <- train_test_list$phens_to_use
-
+}
+    
 ###################################
 # Assign samples
 for (var_assign_subsample in c("sams_for_lik_cross_val", 
@@ -104,19 +107,13 @@ if(Data == "eqtl"){
 }
 dimnames(R.init) <- list(phens_to_use, phens_to_use)
 
-##############################################
-# Initilize Sig list
-Sigl.em.init <- initialize_Sig_list(control = control, 
-                                    Y = Y_zeroed[sams_for_model_training, phens_to_use], 
-                                    nSig = nSig, 
-                                    random_init = grepl("rand", Meth))
-
 
 ##############################################
 # Run MVphen
 if (grepl("MVphen", Meth)) {
   source(file = "scripts/01a_fit_MVphen.R")
 }
+
 
 ##################################################
 # Fit models using XD and/or MASH
